feat (CoreContext): Created new class named CoreContext
This class is in charge of the ownership of BattleCore. This way when a BattleCore instance is modified
Mostly when a save state occurs, it is not necessary to copy the returned BattleCore to other XXXManager instances like
it was done before. Now, you only need to modify the core owned by the context who is shared among all instances.